Split code annotators into separate files
