+++ ExpressionToCodeLib/ExpressionToCodeLib.csproj
@@ -45,6 +45,7 @@
     <Compile Include="CSharpFriendlyTypeName.cs" />
     <Compile Include="EqualityExpressions.cs" />
     <Compile Include="ExpressionCompiler.cs" />
+    <Compile Include="SubExpressionPerLineCodeAnnotator.cs" />
     <Compile Include="Unstable_v2_Api\ExpressionWithName.cs" />
     <Compile Include="Unstable_v2_Api\IObjectToCode.cs" />
     <Compile Include="Unstable_v2_Api\ObjectStringify.cs" />
@@ -67,6 +68,7 @@
     <Compile Include="ReflectionHelpers.cs" />
     <Compile Include="UnitTestingFailure.cs" />
     <Compile Include="Unstable_v2_Api\ExpressionWithValue.cs" />
+    <Compile Include="ValuesOnStalksCodeAnnotator.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="ExpressionToCodeLib.nuspec">
